+++ app/src/main/java/com/gradproj/SpeedoTransferApp/MainActivity.kt
@@ -1,26 +1,20 @@
 package com.gradproj.SpeedoTransferApp
 
 
-import android.content.Context
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Surface
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.unit.dp
-import androidx.navigation.NavController
-import androidx.navigation.compose.rememberNavController
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.gradproj.SpeedoTransferApp.features.navigation.Navigation
-import com.gradproj.SpeedoTransferApp.features.navigation.Screen
-import com.gradproj.SpeedoTransferApp.features.prefrences.PrefrencesHelper.isFirstTimeLaunch
+import androidx.navigation.compose.rememberNavController
+import com.gradproj.SpeedoTransferApp.navigation.Navigation
+
+import com.gradproj.SpeedoTransferApp.prefrences.PrefrencesHelper.isFirstTimeLaunch
 import com.gradproj.SpeedoTransferApp.ui.theme.SpeedoTransferAppTheme
 import com.gradproj.SpeedoTransferApp.ui.theme.redGradient
 import com.gradproj.SpeedoTransferApp.ui.theme.white
