--- app/src/main/java/com/gradproj/SpeedoTransferApp/features/profile/Settings.kt
@@ -1,82 +1,0 @@
-package com.gradproj.SpeedoTransferApp.features.profile
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.navigation.NavController
-import androidx.navigation.compose.rememberNavController
-import com.gradproj.SpeedoTransferApp.R
-import com.gradproj.SpeedoTransferApp.features.navigation.Screen
-import com.gradproj.SpeedoTransferApp.ui.components.ProfileComponent
-
-@Composable
-fun Settings(navController: NavController, modifier: Modifier = Modifier) {
-    Column(
-        verticalArrangement = Arrangement.Center,
-        modifier = Modifier.padding(16.dp)
-    ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = modifier
-                .fillMaxWidth()
-                .padding(bottom = 32.dp)
-        ) {
-            Icon(
-                painter = painterResource(id = R.drawable.drop_downic),
-                contentDescription = "back",
-                modifier = Modifier
-                    .clickable {
-                        navController.popBackStack()
-                    }
-            )
-            Spacer(modifier = Modifier.weight(1f))
-            Text(
-                text = "Settings",
-                fontWeight = FontWeight.W500,
-                fontSize = 20.sp,
-                modifier = Modifier
-                    .padding(end = 24.dp)
-            )
-            Spacer(modifier = Modifier.weight(1f))
-        }
-
-        ProfileComponent(
-            navController,
-            pageToGo = Screen.ChangePassword.route,
-            header = "Change password",
-            description = "Change password",
-            icon = ImageVector.vectorResource(id = R.drawable.lock),
-        )
-
-        ProfileComponent(
-            navController,
-            pageToGo = Screen.EditProfile.route,
-            header = "Edit Profile",
-            description = "Change your information",
-            icon = ImageVector.vectorResource(id = R.drawable.editic),
-            modifier = Modifier.padding(top = 15.dp)
-        )
-    }
-}
-
-@Preview(showBackground = true, showSystemUi = true)
-@Composable
-private fun SettingsPreview() {
-    Settings(navController = rememberNavController())
-}
